+++ libs/wire-api/src/Wire/API/Conversation.hs
@@ -482,13 +482,15 @@ newConvSchema :: ValueSchema NamedSwaggerDoc NewConv
 newConvSchema =
   objectWithDocModifier
     "NewConv"
-    (description ?~ "JSON object to create a new conversation")
+    (description ?~ "JSON object to create a new conversation. When using 'qualified_users' (preferred), you can omit 'users'")
     $ NewConv
       <$> newConvUsers
-        .= fieldWithDocModifier
-          "users"
-          (description ?~ usersDesc)
-          (array schema)
+        .= ( fieldWithDocModifier
+               "users"
+               (description ?~ usersDesc)
+               (array schema)
+               <|> pure []
+           )
       <*> newConvQualifiedUsers
         .= ( fieldWithDocModifier
                "qualified_users"
@@ -518,7 +520,7 @@ newConvSchema =
           )
       <*> newConvReceiptMode .= opt (field "receipt_mode" schema)
       <*> newConvUsersRole
-        .= ( field "conversation_role" schema
+        .= ( fieldWithDocModifier "conversation_role" (description ?~ usersRoleDesc) schema
                <|> pure roleNameWireAdmin
            )
   where
@@ -528,6 +530,14 @@ newConvSchema =
     qualifiedUsersDesc =
       "List of qualified user IDs (excluding the requestor) \
       \to be part of this conversation"
+    usersRoleDesc :: Text
+    usersRoleDesc =
+      cs $
+        "The conversation permissions the users \
+        \added in this request should have. \
+        \Optional, defaults to '"
+          <> show roleNameWireAdmin
+          <> "' if unset."
 
 newConvIsManaged :: NewConv -> Bool
 newConvIsManaged = maybe False cnvManaged . newConvTeam