+++ services/brig/test/integration/Util.hs
@@ -743,7 +743,10 @@ someLastPrekeys =
   ]
 
 defPassword :: PlainTextPassword
-defPassword = PlainTextPassword "secret"
+defPassword = PlainTextPassword defPasswordText
+
+defPasswordText :: Text
+defPasswordText = "secret"
 
 defWrongPassword :: PlainTextPassword
 defWrongPassword = PlainTextPassword "not secret"