+++ services/galley/src/Galley/API/Public.hs
@@ -92,7 +92,7 @@ servantSitemap =
         GalleyAPI.getTeamConversation = Teams.getTeamConversation,
         GalleyAPI.deleteTeamConversation = Teams.deleteTeamConversation,
         GalleyAPI.postOtrMessageUnqualified = Update.postOtrMessageUnqualified,
-        GalleyAPI.postOtrMessage = Update.postOtrMessage
+        GalleyAPI.postProteusMessage = Update.postProteusMessage
       }
 
 sitemap :: Routes ApiBuilder Galley ()