+++ CHANGELOG.md
@@ -16,13 +16,26 @@
 
 # Unreleased
 
+## Internal changes
+
+* schema-profunctor: add `optField` combinator and corresponding documentation (#1621, #1624).
+
+## Documentation
+
+* Fix validation errors in Swagger documentation (#1625).
+
+# 2021-06-23
+
 ## API Changes
 
-* [Federation] Add qualified endpoint for sending messages at `POST /conversations/:domain/:cnv/otr/messages` (#1593).
+* [Federation] Add qualified endpoint for sending messages at `POST /conversations/:domain/:cnv/proteus/messages` (#1593, #1614, #1616, #1620).
 
 ## Features
 
 ## Bug fixes and other updates
+
+* [helm] Allow sending messages upto 40 MB by default (#1614)
+* Fix for https://github.com/wireapp/wire-webapp/security/advisories/GHSA-382j-mmc8-m5rw  (#1613)
 
 ## Documentation
 
@@ -55,7 +68,7 @@ Deploy brig before galley (#1526, #1549)
 * [Federation] expose /conversations/{cnv}/members/v2 for federation backends (#1543)
 
 ## Bug fixes and other updates
-* Fix MIME-type of asset artifacts 
+* Fix MIME-type of asset artifacts
 * Add some missing charts (#1533)
 
 # Internal changes